[fix] 커서 타입을 size_t로 교체 및 변수명 네이밍 통일   - Lexer.h: m_startIdx, m_currentIdx 타입을 int에서 std::size_t로 변경   - Lexer.cpp: while 조건의 불필요한 캐스트 제거   - LexerTest.cpp: 테스트 변수명 token_array → tokenArray로 통일

+++ CodeInterpreter/Lexer.cpp
@@ -3,7 +3,7 @@
 std::vector<Token> Lexer::tokenize(const std::string& source) {
     reset(source);
 
-    while (m_currentIdx < (int)m_source.size()) {
+    while (m_currentIdx < m_source.size()) {
         m_startIdx = m_currentIdx;
         scanToken();
     }
